refactor(call): move target resolution helpers
