+++ docs/context/interface/api.md
@@ -10,6 +10,7 @@ sources:
   - src/treg/application/auth.py
   - src/treg/application/call/idempotency.py
   - src/treg/application/call/intake.py
+  - src/treg/application/call/resolve.py
   - src/treg/application/call/types.py
   - src/treg/application/connect.py
   - src/treg/application/onboard.py
